+++ dc-cudami-server/dc-cudami-server-backend-api/src/main/java/de/digitalcollections/cudami/server/backend/api/repository/UniqueObjectRepository.java
@@ -3,10 +3,7 @@ package de.digitalcollections.cudami.server.backend.api.repository;
 import de.digitalcollections.cudami.server.backend.api.repository.exceptions.RepositoryException;
 import de.digitalcollections.model.UniqueObject;
 import de.digitalcollections.model.list.filtering.Filtering;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.UUID;
+import java.util.*;
 import java.util.stream.Collectors;
 
 public interface UniqueObjectRepository<U extends UniqueObject>
@@ -71,47 +68,111 @@ public interface UniqueObjectRepository<U extends UniqueObject>
   int deleteByUuids(List<UUID> uuidList) throws RepositoryException;
 
   /**
-   * Retrieve one {@code UniqueObject} by given properties in example instance.
+   * Retrieve one {@code UniqueObject}s by given properties in example instance.
    *
-   * @param uniqueObject example instance containing unique property (PK: UUID)
-   * @return the found {@code UniqueObject} or {@code null}
+   * @param uniqueObject example instancescontaining unique property (PK: UUID)
+   * @return The found {@code UniqueObject} or null
    * @throws RepositoryException in case of technical problems
    */
   default U getByExample(U uniqueObject) throws RepositoryException {
     if (uniqueObject == null) {
       throw new IllegalArgumentException("get failed: given object must not be null");
     }
-    return getByUuid(uniqueObject.getUuid());
+    List<U> uniqueObjects = getByExamples(List.of(uniqueObject));
+    return (uniqueObjects.isEmpty() ? null : uniqueObjects.stream().findFirst().orElse(null));
   }
 
   /**
-   * Retrieve one {@code UniqueObject} by given properties in example instance and given filtering.
+   * Retrieve {@code UniqueObject}s by given properties in example instances.
+   *
+   * @param uniqueObjects example instances containing unique property (PK: UUID)
+   * @return List of found {@code UniqueObject}s
+   * @throws RepositoryException in case of technical problems
+   */
+  default List<U> getByExamples(List<U> uniqueObjects) throws RepositoryException {
+    if (uniqueObjects == null) {
+      throw new IllegalArgumentException("get failed: given objects must not be null");
+    }
+    return getByUuids(
+        uniqueObjects.stream()
+            .map(UniqueObject::getUuid)
+            .filter(Objects::nonNull)
+            .collect(Collectors.toList()));
+  }
+
+  /**
+   * Retrieve a {@code UniqueObject} by given properties in example instance and given filtering.
    *
    * @param uniqueObject example instance containing unique property
    * @param filtering filtering params
-   * @return the found {@code UniqueObject} or {@code null}
+   * @return the found {@code UniqueObject} or null
    * @throws RepositoryException in case of problems
    */
   default U getByExampleAndFiltering(U uniqueObject, Filtering filtering)
       throws RepositoryException {
     if (uniqueObject == null || filtering == null) {
-      throw new IllegalArgumentException("get failed: given objects must not be null");
+      throw new IllegalArgumentException("get failed: given object must not be null");
     }
-    return getByUuidAndFiltering(uniqueObject.getUuid(), filtering);
+    List<U> uniqueObjects = getByExamplesAndFiltering(List.of(uniqueObject), filtering);
+    return (uniqueObjects.isEmpty() ? null : uniqueObjects.stream().findFirst().orElse(null));
   }
 
   /**
-   * Retrieve the {@code UniqueObject} with the supplied UUID (PK).
+   * Retrieve {@code UniqueObject}s by given properties in example instances and given filtering.
+   *
+   * @param uniqueObjects example instances containing unique property
+   * @param filtering filtering params
+   * @return List of found {@code UniqueObject}s
+   * @throws RepositoryException in case of problems
+   */
+  default List<U> getByExamplesAndFiltering(List<U> uniqueObjects, Filtering filtering)
+      throws RepositoryException {
+    if (uniqueObjects == null || filtering == null) {
+      throw new IllegalArgumentException("get failed: given objects must not be null");
+    }
+    return getByUuidsAndFiltering(
+        uniqueObjects.stream()
+            .map(UniqueObject::getUuid)
+            .filter(Objects::nonNull)
+            .collect(Collectors.toList()),
+        filtering);
+  }
+
+  /**
+   * Retrieves the {@code UniqueObject} with the supplied UUIDs (PK).
    *
    * @param uniqueObjectUuid UUID of unique object
-   * @return the found {@code UniqueObject} or {@code null}
+   * @return the found {@code UniqueObject} or null
    * @throws RepositoryException
    */
   default U getByUuid(UUID uniqueObjectUuid) throws RepositoryException {
-    return getByUuidAndFiltering(uniqueObjectUuid, null);
+    List<U> uniqueObjects = getByUuids(List.of(uniqueObjectUuid));
+    return (uniqueObjects.isEmpty() ? null : uniqueObjects.stream().findFirst().orElse(null));
   }
 
-  U getByUuidAndFiltering(UUID uniqueObjectUuid, Filtering filtering) throws RepositoryException;
+  /**
+   * Retrieves the {@code UniqueObject}s with the supplied UUIDs (PK).
+   *
+   * @param uniqueObjectUuids UUIDs of unique objects
+   * @return list of found {@code UniqueObject}s
+   * @throws RepositoryException
+   */
+  default List<U> getByUuids(List<UUID> uniqueObjectUuids) throws RepositoryException {
+    return getByUuidsAndFiltering(uniqueObjectUuids, null);
+  }
+
+  List<U> getByUuidsAndFiltering(List<UUID> uniqueObjectUuids, Filtering filtering)
+      throws RepositoryException;
+
+  default U getByUuidAndFiltering(UUID uniqueObjectUuid, Filtering filtering)
+      throws RepositoryException {
+    if (uniqueObjectUuid == null || filtering == null) {
+      throw new IllegalArgumentException(
+          "get failed: given uuid and/or filtering must not be null");
+    }
+    List<U> uniqueObjects = getByUuidsAndFiltering(List.of(uniqueObjectUuid), filtering);
+    return uniqueObjects.stream().findFirst().orElse(null);
+  }
 
   List<U> getRandom(int count) throws RepositoryException;